clisqlclient: split code in smaller files
Release note: None
