+++ pkg/cli/clisqlclient/BUILD.bazel
@@ -9,8 +9,14 @@ go_library(
         "format_table.go",
         "format_value.go",
         "init_conn_error.go",
+        "make_query.go",
+        "parse_bool.go",
+        "row_strings.go",
+        "rows.go",
+        "run_query.go",
         "table_display_format.go",
         "testutils.go",
+        "txn_shim.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/cli/clisqlclient",
     visibility = ["//visibility:public"],
@@ -46,6 +52,8 @@ go_test(
         "format_table_test.go",
         "format_value_test.go",
         "main_test.go",
+        "parse_bool_test.go",
+        "run_query_test.go",
     ],
     embed = [":clisqlclient"],
     deps = [
