+++ pkg/cli/clisqlclient/conn_util.go
@@ -16,7 +16,6 @@ import (
 	"fmt"
 	"io"
 	"net/url"
-	"reflect"
 	"strconv"
 	"strings"
 	"time"
@@ -26,9 +25,7 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/cli/clierror"
 	"github.com/cockroachdb/cockroach/pkg/roachpb"
 	"github.com/cockroachdb/cockroach/pkg/security"
-	"github.com/cockroachdb/cockroach/pkg/sql/parser"
 	"github.com/cockroachdb/cockroach/pkg/sql/sem/tree"
-	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
 	"github.com/cockroachdb/cockroach/pkg/util/version"
 	"github.com/cockroachdb/errors"
 	"github.com/lib/pq"
@@ -99,6 +96,8 @@ type sqlConn struct {
 	// support connection reconfiguration in the SQL shell.
 	verboseTimings *bool
 }
+
+var _ Conn = (*sqlConn)(nil)
 
 // wrapConnError detects TCP EOF errors during the initial SQL handshake.
 // These are translated to a message "perhaps this is not a CockroachDB node"
@@ -483,35 +482,6 @@ func (c *sqlConn) getLastQueryStatistics() (
 		nil
 }
 
-// sqlTxnShim implements the crdb.Tx interface.
-//
-// It exists to support crdb.ExecuteInTxn. Normally, we'd hand crdb.ExecuteInTxn
-// a sql.Txn, but sqlConn predates go1.8's support for multiple result sets and
-// so deals directly with the lib/pq driver. See #14964.
-//
-// TODO(knz): This code is incorrect, see
-// https://github.com/cockroachdb/cockroach/issues/67261
-type sqlTxnShim struct {
-	conn *sqlConn
-}
-
-var _ crdb.Tx = sqlTxnShim{}
-
-func (t sqlTxnShim) Commit(context.Context) error {
-	return t.conn.Exec(`COMMIT`, nil)
-}
-
-func (t sqlTxnShim) Rollback(context.Context) error {
-	return t.conn.Exec(`ROLLBACK`, nil)
-}
-
-func (t sqlTxnShim) Exec(_ context.Context, query string, values ...interface{}) error {
-	if len(values) != 0 {
-		panic("sqlTxnShim.ExecContext must not be called with values")
-	}
-	return t.conn.Exec(query, nil)
-}
-
 // ExecTxn runs fn inside a transaction and retries it as needed.
 // On non-retryable failures, the transaction is aborted and rolled
 // back; on success, the transaction is committed.
@@ -604,87 +574,6 @@ func (c *sqlConn) silentClose() {
 	}
 }
 
-type sqlRowsI interface {
-	driver.RowsColumnTypeScanType
-	driver.RowsColumnTypeDatabaseTypeName
-	Result() driver.Result
-	Tag() string
-
-	// Go 1.8 multiple result set interfaces.
-	// TODO(mjibson): clean this up after 1.8 is released.
-	HasNextResultSet() bool
-	NextResultSet() error
-}
-
-type sqlRows struct {
-	rows sqlRowsI
-	conn *sqlConn
-}
-
-func (r *sqlRows) Columns() []string {
-	return r.rows.Columns()
-}
-
-func (r *sqlRows) Result() driver.Result {
-	return r.rows.Result()
-}
-
-func (r *sqlRows) Tag() string {
-	return r.rows.Tag()
-}
-
-func (r *sqlRows) Close() error {
-	r.conn.flushNotices()
-	err := r.rows.Close()
-	if errors.Is(err, driver.ErrBadConn) {
-		r.conn.reconnecting = true
-		r.conn.silentClose()
-	}
-	return err
-}
-
-// Next implements the Rows interface.
-func (r *sqlRows) Next(values []driver.Value) error {
-	err := r.rows.Next(values)
-	if errors.Is(err, driver.ErrBadConn) {
-		r.conn.reconnecting = true
-		r.conn.silentClose()
-	}
-	for i, v := range values {
-		if b, ok := v.([]byte); ok {
-			values[i] = append([]byte{}, b...)
-		}
-	}
-	// After the first row was received, we want to delay all
-	// further notices until the end of execution.
-	r.conn.delayNotices = true
-	return err
-}
-
-// NextResultSet prepares the next result set for reading.
-func (r *sqlRows) NextResultSet() (bool, error) {
-	if !r.rows.HasNextResultSet() {
-		return false, nil
-	}
-	return true, r.rows.NextResultSet()
-}
-
-func (r *sqlRows) ColumnTypeScanType(index int) reflect.Type {
-	return r.rows.ColumnTypeScanType(index)
-}
-
-func (r *sqlRows) ColumnTypeDatabaseTypeName(index int) string {
-	return r.rows.ColumnTypeDatabaseTypeName(index)
-}
-
-func (r *sqlRows) ColumnTypeNames() []string {
-	colTypes := make([]string, len(r.Columns()))
-	for i := range colTypes {
-		colTypes[i] = r.ColumnTypeDatabaseTypeName(i)
-	}
-	return colTypes
-}
-
 // MakeSQLConn creates a connection object from a connection URL.
 // The booleans passed by reference are parameters that
 // can be changed in the SQL interactive shell after the connection
@@ -732,373 +621,4 @@ func (c *sqlConn) fillPassword() error {
 	c.url = connURL.String()
 	c.passwordMissing = false
 	return nil
-}
-
-// QueryFn is the type of functions produced by MakeQuery.
-type QueryFn func(conn Conn) (rows Rows, isMultiStatementQuery bool, err error)
-
-// MakeQuery encapsulates a SQL query and its parameter into a
-// function that can be applied to a connection object.
-func MakeQuery(query string, parameters ...driver.Value) QueryFn {
-	return func(conn Conn) (Rows, bool, error) {
-		isMultiStatementQuery := parser.HasMultipleStatements(query)
-		// driver.Value is an alias for interface{}, but must adhere to a restricted
-		// set of types when being passed to driver.Queryer.Query (see
-		// driver.IsValue). We use driver.DefaultParameterConverter to perform the
-		// necessary conversion. This is usually taken care of by the sql package,
-		// but we have to do so manually because we're talking directly to the
-		// driver.
-		for i := range parameters {
-			var err error
-			parameters[i], err = driver.DefaultParameterConverter.ConvertValue(parameters[i])
-			if err != nil {
-				return nil, isMultiStatementQuery, err
-			}
-		}
-		rows, err := conn.Query(query, parameters)
-		return rows, isMultiStatementQuery, err
-	}
-}
-
-// RunQuery takes a 'query' with optional 'parameters'.
-// It runs the sql query and returns a list of columns names and a list of rows.
-func RunQuery(conn Conn, fn QueryFn, showMoreChars bool) ([]string, [][]string, error) {
-	rows, _, err := fn(conn)
-	if err != nil {
-		return nil, nil, err
-	}
-
-	defer func() { _ = rows.Close() }()
-	return sqlRowsToStrings(rows, showMoreChars)
-}
-
-// handleCopyError ensures the user is properly informed when they issue
-// a COPY statement somewhere in their input.
-func handleCopyError(conn *sqlConn, err error) error {
-	if !strings.HasPrefix(err.Error(), "pq: unknown response for simple query: 'G'") {
-		return err
-	}
-
-	// The COPY statement has hosed the connection by putting the
-	// protocol in a state that lib/pq cannot understand any more. Reset
-	// it.
-	_ = conn.Close()
-	conn.reconnecting = true
-	return errors.New("woops! COPY has confused this client! Suggestion: use 'psql' for COPY")
-}
-
-// All tags where the RowsAffected value should be reported to
-// the user.
-var tagsWithRowsAffected = map[string]struct{}{
-	"INSERT":    {},
-	"UPDATE":    {},
-	"DELETE":    {},
-	"DROP USER": {},
-	// This one is used with e.g. CREATE TABLE AS (other SELECT
-	// statements have type Rows, not RowsAffected).
-	"SELECT": {},
-}
-
-// RunQueryAndFormatResults takes a 'query' with optional 'parameters'.
-// It runs the sql query and writes output to 'w'.
-func RunQueryAndFormatResults(
-	c Conn, w io.Writer, fn QueryFn, tableDisplayFormat TableDisplayFormat, tableBorderMode int,
-) (err error) {
-	conn := c.(*sqlConn)
-	startTime := timeutil.Now()
-	rows, isMultiStatementQuery, err := fn(conn)
-	if err != nil {
-		return handleCopyError(conn, err)
-	}
-	defer func() {
-		closeErr := rows.Close()
-		err = errors.CombineErrors(err, closeErr)
-	}()
-	for {
-		// lib/pq is not able to tell us before the first call to Next()
-		// whether a statement returns either
-		// - a rows result set with zero rows (e.g. SELECT on an empty table), or
-		// - no rows result set, but a valid value for RowsAffected (e.g. INSERT), or
-		// - doesn't return any rows whatsoever (e.g. SET).
-		//
-		// To distinguish them we must go through Next() somehow, which is what the
-		// render() function does. So we ask render() to call this noRowsHook
-		// when Next() has completed its work and no rows where observed, to decide
-		// what to do.
-		noRowsHook := func() (bool, error) {
-			res := rows.Result()
-			if ra, ok := res.(driver.RowsAffected); ok {
-				nRows, err := ra.RowsAffected()
-				if err != nil {
-					return false, err
-				}
-
-				// This may be either something like INSERT with a valid
-				// RowsAffected value, or a statement like SET. The pq driver
-				// uses both driver.RowsAffected for both.  So we need to be a
-				// little more manual.
-				tag := rows.Tag()
-				if tag == "SELECT" && nRows == 0 {
-					// As explained above, the pq driver unhelpfully does not
-					// distinguish between a statement returning zero rows and a
-					// statement returning an affected row count of zero.
-					// noRowsHook is called non-discriminatingly for both
-					// situations.
-					//
-					// TODO(knz): meanwhile, there are rare, non-SELECT
-					// statements that have tag "SELECT" but are legitimately of
-					// type RowsAffected. CREATE TABLE AS is one. pq's inability
-					// to distinguish those two cases means that any non-SELECT
-					// statement that legitimately returns 0 rows affected, and
-					// for which the user would expect to see "SELECT 0", will
-					// be incorrectly displayed as an empty row result set
-					// instead. This needs to be addressed by ensuring pq can
-					// distinguish the two cases, or switching to an entirely
-					// different driver altogether.
-					//
-					return false, nil
-				} else if _, ok := tagsWithRowsAffected[tag]; ok {
-					// INSERT, DELETE, etc.: print the row count.
-					nRows, err := ra.RowsAffected()
-					if err != nil {
-						return false, err
-					}
-					fmt.Fprintf(w, "%s %d\n", tag, nRows)
-				} else {
-					// SET, etc.: just print the tag, or OK if there's no tag.
-					if tag == "" {
-						tag = "OK"
-					}
-					fmt.Fprintln(w, tag)
-				}
-				return true, nil
-			}
-			// Other cases: this is a statement with a rows result set, but
-			// zero rows (e.g. SELECT on empty table). Let the reporter
-			// handle it.
-			return false, nil
-		}
-
-		cols := getColumnStrings(rows, true)
-		reporter, cleanup, err := makeReporter(w, tableDisplayFormat, tableBorderMode)
-		if err != nil {
-			return err
-		}
-
-		var queryCompleteTime time.Time
-		completedHook := func() { queryCompleteTime = timeutil.Now() }
-
-		if err := func() error {
-			if cleanup != nil {
-				defer cleanup()
-			}
-			return render(reporter, w, cols, newRowIter(rows, true), completedHook, noRowsHook)
-		}(); err != nil {
-			return err
-		}
-
-		maybeShowTimes(conn, w, isMultiStatementQuery, startTime, queryCompleteTime)
-
-		if more, err := rows.NextResultSet(); err != nil {
-			return err
-		} else if !more {
-			return nil
-		}
-	}
-}
-
-// maybeShowTimes displays the execution time if show_times has been set.
-func maybeShowTimes(
-	conn *sqlConn, w io.Writer, isMultiStatementQuery bool, startTime,
-	queryCompleteTime time.Time,
-) {
-	if !*conn.showTimes {
-		return
-	}
-
-	defer func() {
-		// If there was noticeable overhead, let the user know.
-		renderDelay := timeutil.Now().Sub(queryCompleteTime)
-		if renderDelay >= 1*time.Second && conn.isInteractive {
-			fmt.Fprintf(stderr,
-				"\nNote: an additional delay of %s was spent formatting the results.\n"+
-					"You can use \\set display_format to change the formatting.\n",
-				renderDelay)
-		}
-		// An additional empty line as separator.
-		fmt.Fprintln(w)
-	}()
-
-	clientSideQueryLatency := queryCompleteTime.Sub(startTime)
-	// We don't print timings for multi-statement queries as we don't have an
-	// accurate way to measure them currently. See #48180.
-	if isMultiStatementQuery {
-		// No need to print if no one's watching.
-		if conn.isInteractive {
-			fmt.Fprintf(stderr, "\nNote: timings for multiple statements on a single line are not supported. See %s.\n",
-				build.MakeIssueURL(48180))
-		}
-		return
-	}
-
-	// Print a newline early. This provides a discreet visual
-	// feedback that execution finished, and that the next line of
-	// output will be a warning or execution time(s).
-	fmt.Fprintln(w)
-
-	// We accumulate the timing details into a buffer prior to emitting
-	// them to the output stream, so as to avoid interleaving warnings
-	// or SQL notices with the full timing string.
-	var stats strings.Builder
-
-	// Print a newline so that there is a visual separation between a notice and
-	// the timing information.
-	fmt.Fprintln(&stats)
-
-	// Suggested by Radu: for sub-second results, show simplified
-	// timings in milliseconds.
-	unit := "s"
-	multiplier := 1.
-	precision := 3
-	if clientSideQueryLatency.Seconds() < 1 {
-		unit = "ms"
-		multiplier = 1000.
-		precision = 0
-	}
-
-	if *conn.verboseTimings {
-		fmt.Fprintf(&stats, "Time: %s", clientSideQueryLatency)
-	} else {
-		// Simplified displays: human users typically can't
-		// distinguish sub-millisecond latencies.
-		fmt.Fprintf(&stats, "Time: %.*f%s", precision, clientSideQueryLatency.Seconds()*multiplier, unit)
-	}
-
-	if !*conn.enableServerExecutionTimings {
-		fmt.Fprintln(w, stats.String())
-		return
-	}
-
-	// If discrete server/network timings are available, also print them.
-	parseLat, planLat, execLat, serviceLat, jobsLat, containsJobLat, err := conn.getLastQueryStatistics()
-	if err != nil {
-		fmt.Fprint(w, stats.String())
-		fmt.Fprintf(stderr, "\nwarning: %v", err)
-		return
-	}
-
-	fmt.Fprint(&stats, " total")
-
-	networkLat := clientSideQueryLatency - (serviceLat + jobsLat)
-	// serviceLat can be greater than clientSideQueryLatency for some extremely quick
-	// statements (eg. BEGIN). So as to not confuse the user, we attribute all of
-	// the clientSideQueryLatency to the network in such cases.
-	if networkLat.Seconds() < 0 {
-		networkLat = clientSideQueryLatency
-	}
-	otherLat := serviceLat - parseLat - planLat - execLat
-	if *conn.verboseTimings {
-		// Only display schema change latency if the server provided that
-		// information to not confuse users.
-		// TODO(arul): this can be removed in 22.1.
-		if containsJobLat {
-			fmt.Fprintf(&stats, " (parse %s / plan %s / exec %s / schema change %s / other %s / network %s)",
-				parseLat, planLat, execLat, jobsLat, otherLat, networkLat)
-		} else {
-			fmt.Fprintf(&stats, " (parse %s / plan %s / exec %s / other %s / network %s)",
-				parseLat, planLat, execLat, otherLat, networkLat)
-		}
-	} else {
-		// Simplified display: just show the execution/network breakdown.
-		//
-		// Note: we omit the report details for queries that
-		// last for a millisecond or less. This is because for such
-		// small queries, the detail is just noise to the human observer.
-		sep := " ("
-		reportTiming := func(label string, lat time.Duration) {
-			fmt.Fprintf(&stats, "%s%s %.*f%s", sep, label, precision, lat.Seconds()*multiplier, unit)
-			sep = " / "
-		}
-		reportTiming("execution", serviceLat+jobsLat)
-		reportTiming("network", networkLat)
-		fmt.Fprint(&stats, ")")
-	}
-	fmt.Fprintln(w, stats.String())
-}
-
-// sqlRowsToStrings turns 'rows' into a list of rows, each of which
-// is a list of column values.
-// 'rows' should be closed by the caller.
-// It returns the header row followed by all data rows.
-// If both the header row and list of rows are empty, it means no row
-// information was returned (eg: statement was not a query).
-// If showMoreChars is true, then more characters are not escaped.
-func sqlRowsToStrings(rows Rows, showMoreChars bool) ([]string, [][]string, error) {
-	cols := getColumnStrings(rows, showMoreChars)
-	allRows, err := getAllRowStrings(rows, rows.ColumnTypeNames(), showMoreChars)
-	if err != nil {
-		return nil, nil, err
-	}
-	return cols, allRows, nil
-}
-
-func getColumnStrings(rows Rows, showMoreChars bool) []string {
-	srcCols := rows.Columns()
-	cols := make([]string, len(srcCols))
-	for i, c := range srcCols {
-		cols[i] = FormatVal(c, "NAME", showMoreChars, showMoreChars)
-	}
-	return cols
-}
-
-func getAllRowStrings(rows Rows, colTypes []string, showMoreChars bool) ([][]string, error) {
-	var allRows [][]string
-
-	for {
-		rowStrings, err := getNextRowStrings(rows, colTypes, showMoreChars)
-		if err != nil {
-			return nil, err
-		}
-		if rowStrings == nil {
-			break
-		}
-		allRows = append(allRows, rowStrings)
-	}
-
-	return allRows, nil
-}
-
-func getNextRowStrings(rows Rows, colTypes []string, showMoreChars bool) ([]string, error) {
-	cols := rows.Columns()
-	var vals []driver.Value
-	if len(cols) > 0 {
-		vals = make([]driver.Value, len(cols))
-	}
-
-	err := rows.Next(vals)
-	if err == io.EOF {
-		return nil, nil
-	}
-	if err != nil {
-		return nil, err
-	}
-
-	rowStrings := make([]string, len(cols))
-	for i, v := range vals {
-		rowStrings[i] = FormatVal(v, colTypes[i], showMoreChars, showMoreChars)
-	}
-	return rowStrings, nil
-}
-
-// ParseBool parses a boolean string for use in CLI SQL commands.
-// It recognizes booleans in a similar way as 'psql'.
-func ParseBool(s string) (bool, error) {
-	switch strings.TrimSpace(strings.ToLower(s)) {
-	case "true", "on", "yes", "1":
-		return true, nil
-	case "false", "off", "no", "0":
-		return false, nil
-	default:
-		return false, errors.Newf("invalid boolean value %q", s)
-	}
 }
