+++ pkg/cli/clisqlclient/conn_util_test.go
@@ -11,12 +11,9 @@
 package clisqlclient_test
 
 import (
-	"bytes"
 	"database/sql/driver"
 	"fmt"
-	"io"
 	"net/url"
-	"reflect"
 	"testing"
 
 	"github.com/cockroachdb/cockroach/pkg/cli"
@@ -26,8 +23,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/testutils/sqlutils"
 	"github.com/cockroachdb/cockroach/pkg/util/leaktest"
 	"github.com/cockroachdb/errors"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func makeSQLConn(url string) clisqlclient.Conn {
@@ -39,15 +34,6 @@ func makeSQLConn(url string) clisqlclient.Conn {
 		&bf,   /* echo */
 		&bf,   /* showTimes */
 		&bf,   /* verboseTimings */
-	)
-}
-
-func runQueryAndFormatResults(
-	conn clisqlclient.Conn, w io.Writer, fn clisqlclient.QueryFn,
-) (err error) {
-	return clisqlclient.RunQueryAndFormatResults(conn, w, fn,
-		clisqlclient.TableDisplayTable,
-		0, /* tableBorderMode */
 	)
 }
 
@@ -133,189 +119,6 @@ func simulateServerRestart(
 	return cleanup2
 }
 
-func TestRunQuery(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	c := cli.NewCLITest(cli.TestCLIParams{T: t})
-	defer c.Cleanup()
-
-	url, cleanup := sqlutils.PGUrl(t, c.ServingSQLAddr(), t.Name(), url.User(security.RootUser))
-	defer cleanup()
-
-	conn := makeSQLConn(url.String())
-	defer func() {
-		if err := conn.Close(); err != nil {
-			t.Fatal(err)
-		}
-	}()
-
-	var b bytes.Buffer
-
-	// Non-query statement.
-	if err := runQueryAndFormatResults(conn, &b,
-		clisqlclient.MakeQuery(`SET DATABASE=system`)); err != nil {
-		t.Fatal(err)
-	}
-
-	expected := `
-SET
-`
-	if a, e := b.String(), expected[1:]; a != e {
-		t.Fatalf("expected output:\n%s\ngot:\n%s", e, a)
-	}
-	b.Reset()
-
-	// Use system database for sample query/output as they are fairly fixed.
-	cols, rows, err := clisqlclient.RunQuery(conn, clisqlclient.MakeQuery(`SHOW COLUMNS FROM system.namespace`), false)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	expectedCols := []string{
-		"column_name",
-		"data_type",
-		"is_nullable",
-		"column_default",
-		"generation_expression",
-		"indices",
-		"is_hidden",
-	}
-	if !reflect.DeepEqual(expectedCols, cols) {
-		t.Fatalf("expected:\n%v\ngot:\n%v", expectedCols, cols)
-	}
-
-	expectedRows := [][]string{
-		{`parentID`, `INT8`, `false`, `NULL`, ``, `{primary}`, `false`},
-		{`parentSchemaID`, `INT8`, `false`, `NULL`, ``, `{primary}`, `false`},
-		{`name`, `STRING`, `false`, `NULL`, ``, `{primary}`, `false`},
-		{`id`, `INT8`, `true`, `NULL`, ``, `{primary}`, `false`},
-	}
-	if !reflect.DeepEqual(expectedRows, rows) {
-		t.Fatalf("expected:\n%v\ngot:\n%v", expectedRows, rows)
-	}
-
-	if err := runQueryAndFormatResults(conn, &b,
-		clisqlclient.MakeQuery(`SHOW COLUMNS FROM system.namespace`)); err != nil {
-		t.Fatal(err)
-	}
-
-	expected = `
-   column_name   | data_type | is_nullable | column_default | generation_expression |  indices  | is_hidden
------------------+-----------+-------------+----------------+-----------------------+-----------+------------
-  parentID       | INT8      |    false    | NULL           |                       | {primary} |   false
-  parentSchemaID | INT8      |    false    | NULL           |                       | {primary} |   false
-  name           | STRING    |    false    | NULL           |                       | {primary} |   false
-  id             | INT8      |    true     | NULL           |                       | {primary} |   false
-(4 rows)
-`
-
-	if a, e := b.String(), expected[1:]; a != e {
-		t.Fatalf("expected output:\n%s\ngot:\n%s", e, a)
-	}
-	b.Reset()
-
-	// Test placeholders.
-	if err := runQueryAndFormatResults(conn, &b,
-		clisqlclient.MakeQuery(`SELECT * FROM system.namespace WHERE name=$1`, "descriptor")); err != nil {
-		t.Fatal(err)
-	}
-
-	expected = `
-  parentID | parentSchemaID |    name    | id
------------+----------------+------------+-----
-         1 |             29 | descriptor |  3
-(1 row)
-`
-	if a, e := b.String(), expected[1:]; a != e {
-		t.Fatalf("expected output:\n%s\ngot:\n%s", e, a)
-	}
-	b.Reset()
-
-	// Test multiple results.
-	if err := runQueryAndFormatResults(conn, &b,
-		clisqlclient.MakeQuery(`SELECT 1 AS "1"; SELECT 2 AS "2", 3 AS "3"; SELECT 'hello' AS "'hello'"`)); err != nil {
-		t.Fatal(err)
-	}
-
-	expected = `
-  1
------
-  1
-(1 row)
-  2 | 3
-----+----
-  2 | 3
-(1 row)
-  'hello'
------------
-  hello
-(1 row)
-`
-
-	if a, e := b.String(), expected[1:]; a != e {
-		t.Fatalf("expected output:\n%s\ngot:\n%s", e, a)
-	}
-	b.Reset()
-}
-
-func TestUtfName(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-
-	c := cli.NewCLITest(cli.TestCLIParams{T: t})
-	defer c.Cleanup()
-
-	url, cleanup := sqlutils.PGUrl(t, c.ServingSQLAddr(), t.Name(), url.User(security.RootUser))
-	defer cleanup()
-
-	conn := makeSQLConn(url.String())
-	defer func() {
-		if err := conn.Close(); err != nil {
-			t.Fatal(err)
-		}
-	}()
-
-	var b bytes.Buffer
-
-	if err := runQueryAndFormatResults(conn, &b,
-		clisqlclient.MakeQuery(`CREATE DATABASE test_utf;
-CREATE TABLE test_utf.żółw (id INT PRIMARY KEY, value INT);
-ALTER TABLE test_utf.żółw ADD CONSTRAINT żó UNIQUE (value)`)); err != nil {
-		t.Fatal(err)
-	}
-
-	b.Reset()
-	if err := runQueryAndFormatResults(conn, &b,
-		clisqlclient.MakeQuery(`SHOW TABLES FROM test_utf;`)); err != nil {
-		t.Fatal(err)
-	}
-	expected := `
-  schema_name | table_name | type  | owner | estimated_row_count | locality
---------------+------------+-------+-------+---------------------+-----------
-  public      | żółw       | table | root  |                NULL | NULL
-(1 row)
-`
-	if a, e := b.String(), expected[1:]; a != e {
-		t.Errorf("expected output:\n%s\ngot:\n%s", e, a)
-	}
-	b.Reset()
-
-	if err := runQueryAndFormatResults(conn, &b,
-		clisqlclient.MakeQuery(`SHOW CONSTRAINTS FROM test_utf.żółw;`)); err != nil {
-		t.Fatal(err)
-	}
-	expected = `
-  table_name | constraint_name | constraint_type |       details        | validated
--------------+-----------------+-----------------+----------------------+------------
-  żółw       | primary         | PRIMARY KEY     | PRIMARY KEY (id ASC) |   true
-  żółw       | żó              | UNIQUE          | UNIQUE (value ASC)   |   true
-(2 rows)
-`
-	if a, e := b.String(), expected[1:]; a != e {
-		t.Errorf("expected output:\n%s\ngot:\n%s", e, a)
-	}
-	b.Reset()
-}
-
 func TestTransactionRetry(t *testing.T) {
 	defer leaktest.AfterTest(t)()
 
@@ -361,41 +164,5 @@ func TestTransactionRetry(t *testing.T) {
 	}
 	if tries <= 2 {
 		t.Fatalf("expected transaction to require at least two tries, but it only required %d", tries)
-	}
-}
-
-func TestParseBool(t *testing.T) {
-	defer leaktest.AfterTest(t)()
-	testcases := []struct {
-		input     string
-		expect    bool
-		expectErr bool
-	}{
-		{"true", true, false},
-		{"on", true, false},
-		{"yes", true, false},
-		{"1", true, false},
-		{" TrUe	", true, false},
-
-		{"false", false, false},
-		{"off", false, false},
-		{"no", false, false},
-		{"0", false, false},
-		{"	FaLsE ", false, false},
-
-		{"", false, true},
-		{"foo", false, true},
-	}
-
-	for _, tc := range testcases {
-		t.Run(tc.input, func(t *testing.T) {
-			b, err := clisqlclient.ParseBool(tc.input)
-			if tc.expectErr {
-				require.Error(t, err)
-			} else {
-				require.NoError(t, err)
-				assert.Equal(t, tc.expect, b)
-			}
-		})
 	}
 }
